refactor: vendor frontend dependencies and remove CDN links
Replace external CDN links with locally hosted Bootstrap, FontAwesome, and Vue libraries to improve reliability and enable offline functionality.

Key changes:
- Add @fortawesome/fontawesome-free, bootstrap, vue, and sharp as dependencies
- Create copy-vendor script to extract necessary files to src/public/vendor/
- Update all EJS templates to reference local vendor assets instead of CDNs
- Modify Dockerfile to run copy-vendor during build process
- Add src/public/vendor/ to .gitignore to exclude copied files
- Include scripts/ directory in Docker builds
- Add VSCode launch configuration for Bun debugging

This change reduces external dependencies, improves load times, and ensures the dashboard remains functional without internet access.
