+++ .dockerignore
@@ -8,9 +8,6 @@ data/
 # Development and test data
 data/test/
 
-# Scripts
-scripts/
-
 # Environment files
 .env
 .env.*
