+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/
 data-test/
 .env
 data/
+src/public/vendor/
