+++ .vscode/launch.json
@@ -1,0 +1,20 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        //debug with bun
+        {
+            "type": "bun",
+            "request": "launch",
+            "name": "Debug with Bun",
+            "runtimeArgs": [
+                "--inspect-brk",
+                "--watch",
+                "src/index.ts"
+            ],
+        }
+   
+    ]
+}
