+++ docker-compose.dev.yml
@@ -15,5 +15,4 @@ services:
     volumes:
       - ./data:/app/data
       - /var/run/docker.sock:/var/run/docker.sock
-      - ./:/app
     restart: unless-stopped
