+++ src/lib/dashboardSettings.ts
@@ -2,21 +2,10 @@ import fs from "fs/promises";
 import path from "path";
 import { resolveDataPath } from "./dataPaths.js";
 import { isDemoMode, logDemoAction } from "./demoMode.js";
-import { validateOrThrow, updateSettingsSchema } from "./validation.js";
 
-export const DEFAULT_DASHBOARD_SETTINGS = {
-  appTitle: "Leet Container Dashboard",
-  appSlogan: "Monitor and control containers on your network.",
-  theme: "dark",
-  backgroundImagePath: "",
-  hideAttributionFooter: false,
-  showContainerResources: true,
-  showServerResources: true,
-  showImageName: true,
-  showContainerHash: true,
-} as const;
+const DEFAULT_DASHBOARD_SETTINGS_PATH = resolveDataPath("dashboardSettings.json");
 
-export type DashboardTheme = "light" | "dark";
+export type DashboardTheme = "dark" | "light";
 
 export type DashboardSettings = {
   appTitle: string;
@@ -30,55 +19,34 @@ export type DashboardSettings = {
   showContainerHash: boolean;
 };
 
-const DEFAULT_DASHBOARD_SETTINGS_PATH = resolveDataPath("dashboardSettings.json");
-const DEFAULT_DASHBOARD_UPLOADS_PATH = resolveDataPath("uploads", "backgrounds");
+export const DEFAULT_DASHBOARD_SETTINGS: DashboardSettings = {
+  appTitle: "Leet Container Dashboard",
+  appSlogan: "Manage your containers with style",
+  theme: "dark",
+  backgroundImagePath: "",
+  hideAttributionFooter: false,
+  showContainerResources: true,
+  showServerResources: true,
+  showImageName: true,
+  showContainerHash: true,
+};
 
-export function getDashboardSettingsFilePath(): string {
+function getDashboardSettingsFilePath(): string {
   return process.env.DASHBOARD_SETTINGS_FILE || DEFAULT_DASHBOARD_SETTINGS_PATH;
 }
 
-export function getDashboardBackgroundUploadsPath(): string {
-  return process.env.DASHBOARD_UPLOADS_DIR || DEFAULT_DASHBOARD_UPLOADS_PATH;
-}
-
-function toBackgroundImagePath(value: unknown): string {
-  if (typeof value !== "string") {
-    return "";
-  }
-
-  const normalized = value.trim();
-  if (!normalized.startsWith("/uploads/backgrounds/")) {
-    return "";
-  }
-
-  return normalized;
-}
-
-function normalizeSettings(input: Partial<DashboardSettings>): DashboardSettings {
-  // Use Joi validation for the core settings
-  const validated = validateOrThrow<{
-    appTitle: string;
-    appSlogan: string;
-    theme: DashboardTheme;
-    hideAttributionFooter: boolean;
-    showContainerResources: boolean;
-    showServerResources: boolean;
-    showImageName: boolean;
-    showContainerHash: boolean;
-  }>(updateSettingsSchema, {
-    appTitle: input.appTitle,
-    appSlogan: input.appSlogan,
-    theme: input.theme,
-    hideAttributionFooter: input.hideAttributionFooter,
-    showContainerResources: input.showContainerResources,
-    showServerResources: input.showServerResources,
-    showImageName: input.showImageName,
-    showContainerHash: input.showContainerHash,
-  });
-
+function validateDashboardSettings(input: unknown): DashboardSettings {
+  const parsed = input as Partial<DashboardSettings>;
   return {
-    ...validated,
-    backgroundImagePath: toBackgroundImagePath(input.backgroundImagePath),
+    appTitle: typeof parsed?.appTitle === "string" ? parsed.appTitle : DEFAULT_DASHBOARD_SETTINGS.appTitle,
+    appSlogan: typeof parsed?.appSlogan === "string" ? parsed.appSlogan : DEFAULT_DASHBOARD_SETTINGS.appSlogan,
+    theme: parsed?.theme === "light" ? "light" : "dark",
+    backgroundImagePath: typeof parsed?.backgroundImagePath === "string" ? parsed.backgroundImagePath : "",
+    hideAttributionFooter: typeof parsed?.hideAttributionFooter === "boolean" ? parsed.hideAttributionFooter : DEFAULT_DASHBOARD_SETTINGS.hideAttributionFooter,
+    showContainerResources: typeof parsed?.showContainerResources === "boolean" ? parsed.showContainerResources : DEFAULT_DASHBOARD_SETTINGS.showContainerResources,
+    showServerResources: typeof parsed?.showServerResources === "boolean" ? parsed.showServerResources : DEFAULT_DASHBOARD_SETTINGS.showServerResources,
+    showImageName: typeof parsed?.showImageName === "boolean" ? parsed.showImageName : DEFAULT_DASHBOARD_SETTINGS.showImageName,
+    showContainerHash: typeof parsed?.showContainerHash === "boolean" ? parsed.showContainerHash : DEFAULT_DASHBOARD_SETTINGS.showContainerHash,
   };
 }
 
@@ -86,49 +54,44 @@ async function ensureDashboardSettingsFileExists(filePath = getDashboardSettings
   try {
     await fs.access(filePath);
   } catch {
+    const defaultSettings = { ...DEFAULT_DASHBOARD_SETTINGS };
     await fs.mkdir(path.dirname(filePath), { recursive: true });
-    await fs.writeFile(filePath, JSON.stringify(DEFAULT_DASHBOARD_SETTINGS, null, 2), "utf-8");
+    await fs.writeFile(filePath, JSON.stringify(defaultSettings, null, 2), "utf-8");
   }
 }
 
-export async function getDashboardSettings(filePath = getDashboardSettingsFilePath()): Promise<DashboardSettings> {
+async function readDashboardSettings(filePath = getDashboardSettingsFilePath()): Promise<DashboardSettings> {
   await ensureDashboardSettingsFileExists(filePath);
-
-  try {
-    const fileContent = await fs.readFile(filePath, "utf-8");
-    const parsed = JSON.parse(fileContent) as Partial<DashboardSettings>;
-    const normalized = normalizeSettings(parsed);
-
-    if (JSON.stringify(parsed) !== JSON.stringify(normalized)) {
-      await saveDashboardSettings(normalized, filePath);
-    }
-
-    return normalized;
-  } catch {
-    await saveDashboardSettings(DEFAULT_DASHBOARD_SETTINGS, filePath);
-    return {
-      ...DEFAULT_DASHBOARD_SETTINGS,
-    };
-  }
+  const text = await fs.readFile(filePath, "utf-8");
+  const parsed = JSON.parse(text) as unknown;
+  return validateDashboardSettings(parsed);
 }
 
-export async function saveDashboardSettings(settings: Partial<DashboardSettings>, filePath = getDashboardSettingsFilePath()): Promise<DashboardSettings> {
-  const normalized = normalizeSettings(settings);
-  
+async function writeDashboardSettings(settings: DashboardSettings, filePath = getDashboardSettingsFilePath()): Promise<void> {
+  const validated = validateDashboardSettings(settings);
+
   if (isDemoMode()) {
-    logDemoAction("saveDashboardSettings", { settings: normalized });
-    return normalized;
+    logDemoAction("writeDashboardSettings", validated);
+    return;
   }
-  
-  await fs.mkdir(path.dirname(filePath), { recursive: true });
-  await fs.writeFile(filePath, JSON.stringify(normalized, null, 2), "utf-8");
-  return normalized;
+
+  await fs.writeFile(filePath, JSON.stringify(validated, null, 2), "utf-8");
 }
 
-export async function updateDashboardSettings(
-  patch: Partial<DashboardSettings>,
-  filePath = getDashboardSettingsFilePath()
-): Promise<DashboardSettings> {
-  const current = await getDashboardSettings(filePath);
-  return saveDashboardSettings({ ...current, ...patch }, filePath);
+export async function getDashboardSettings(): Promise<DashboardSettings> {
+  return readDashboardSettings();
+}
+
+export async function updateDashboardSettings(patch: Partial<DashboardSettings>): Promise<void> {
+  const current = await readDashboardSettings();
+  const updated: DashboardSettings = { ...current, ...patch };
+  await writeDashboardSettings(updated);
+}
+
+export function getDashboardBackgroundUploadsPath(): string {
+  return resolveDataPath("uploads", "backgrounds");
+}
+
+export function getLaunchpadIconsUploadsPath(): string {
+  return resolveDataPath("uploads", "launchpad-icons");
 }
