--- src/views/launcher.ejs
@@ -1,117 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title><%= dashboardSettings.appTitle %> - Launcher</title>
-  <link rel="preconnect" href="https://fonts.googleapis.com" />
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-  <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap" rel="stylesheet" />
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" />
-  <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css" rel="stylesheet" />
-  <link href="/public/styles.css" rel="stylesheet" />
-</head>
-<body class="<%= pageBodyClass %>" data-hs-bg-image="<%= pageBackgroundImagePath %>">
-  <%- include('partials/demoBanner') %>
-
-  <div id="app">
-    <div class="container py-4">
-      <div class="panel mb-4">
-        <div class="d-flex justify-content-between align-items-start gap-3 flex-wrap">
-          <div>
-            <h1 class="mb-2"><i class="fa-solid fa-rocket me-2"></i>Service Launcher</h1>
-            <p class="text-muted mb-0">Jump straight into your web services.</p>
-          </div>
-          <nav class="hs-page-nav" aria-label="Primary navigation">
-            <a href="/" class="btn btn-sm btn-hs-secondary">
-              <i class="fa-solid fa-table-list me-1"></i>Dashboard
-            </a>
-            <a href="/launcher" class="btn btn-sm btn-hs-secondary active" aria-current="page">
-              <i class="fa-solid fa-rocket me-1"></i>Launcher
-            </a>
-          </nav>
-        </div>
-      </div>
-
-      <% if (notice) { %>
-        <div class="alert alert-success hs-flash-alert"><i class="fa-solid fa-circle-check me-2"></i><%= notice %></div>
-      <% } %>
-      <% if (error) { %>
-        <div class="alert alert-danger hs-flash-alert"><i class="fa-solid fa-triangle-exclamation me-2"></i><%= error %></div>
-      <% } %>
-
-      <div class="panel mb-3">
-        <div class="d-flex justify-content-between align-items-center gap-2 flex-wrap">
-          <div class="text-muted small">Showing services from all enabled servers</div>
-          <div class="form-check form-switch launcher-toggle">
-            <input class="form-check-input" type="checkbox" id="showAllToggle" v-model="showAll" />
-            <label class="form-check-label" for="showAllToggle">Show all</label>
-          </div>
-        </div>
-      </div>
-
-      <div v-if="loading" class="text-center py-5">
-        <div class="spinner-border text-primary" role="status"><span class="visually-hidden">Loading...</span></div>
-      </div>
-
-      <div v-else-if="filteredTiles.length" class="launcher-grid">
-        <a v-for="tile in filteredTiles" :key="tile.id" :href="tile.launchUrl" target="_blank" rel="noopener noreferrer" class="launcher-tile">
-          <div :class="['launcher-tile-icon', tile.iconColorClass || 'launcher-icon-default']">
-            <i :class="tile.iconClass || 'fa-solid fa-rocket'"></i>
-          </div>
-          <div class="launcher-tile-content">
-            <h2 class="h6 mb-1">{{ tile.name }}</h2>
-            <p class="text-muted small mb-1 text-break">{{ tile.description }}</p>
-            <span class="launcher-url">{{ tile.publicUrl || tile.localUrl }}</span>
-          </div>
-        </a>
-      </div>
-
-      <div v-else class="panel text-center text-muted py-5">
-        No launchable HTTP services found.
-      </div>
-    </div>
-  </div>
-
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/vue@3.4.21/dist/vue.global.prod.js"></script>
-  <%- include('partials/footer') %>
-  <script>
-    const { createApp } = Vue;
-
-    createApp({
-      data() {
-        return {
-          loading: true,
-          launcherTiles: [],
-          showAll: false,
-        };
-      },
-      computed: {
-        filteredTiles() {
-          if (this.showAll) {
-            return this.launcherTiles;
-          }
-          return this.launcherTiles.filter(tile => !tile.hidden);
-        }
-      },
-      async mounted() {
-        try {
-          const response = await fetch('/api/launcher', { credentials: 'same-origin' });
-          if (!response.ok) {
-            return;
-          }
-          const result = await response.json();
-          if (result.success && result.data) {
-            this.launcherTiles = result.data.launcherTiles || [];
-          }
-        } catch (error) {
-          console.error('Failed to load launcher tiles', error);
-        } finally {
-          this.loading = false;
-        }
-      }
-    }).mount('#app');
-  </script>
-</body>
-</html>
