+++ src/views/users.ejs
@@ -4,11 +4,12 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>User Management - <%= dashboardSettings.appTitle %></title>
+  <link rel="icon" href="/public/logo.png" type="image/png" />
   <link rel="preconnect" href="https://fonts.googleapis.com" />
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
   <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap" rel="stylesheet" />
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" />
-  <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css" rel="stylesheet" />
+  <link href="/public/vendor/bootstrap/css/bootstrap.min.css" rel="stylesheet" />
+  <link href="/public/vendor/fontawesome/css/all.min.css" rel="stylesheet" />
   <link href="/public/styles.css" rel="stylesheet" />
 </head>
 <body class="<%= pageBodyClass %>" data-hs-bg-image="<%= pageBackgroundImagePath %>">
@@ -251,7 +252,7 @@
 
   <%- include('partials/footer') %>
 
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
+  <script src="/public/vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
   <script>
     (() => {
       const configuredBackground = document.body.getAttribute('data-hs-bg-image') || '';
